refactor: Implement Savings Goal feature and unify Category model
This commit introduces the Savings Goal feature, refactors the data model to use a unified `CategoryEntity` for all transaction types, and enhances the UI for detailed savings tracking.

- **Data Model & Repository:**
    - Replaces the specific `SavingCategoryEntity` with a more generic `CategoryEntity` that now includes a `type` field (`income`, `expense`, `saving`).
    - Removes `SavingCategoryEntity` and its generated file.
    - Introduces `SavingEntity` to store savings goal details like `target_amount`.
    - Updates `CategoryRepository` to fetch categories by `TypeTransaction` (`getCategoriesByType`).
    - Extends `SummaryRepository` with `getTotalAmountByCategoryId` and `getTargetAmountByCategoryId` to fetch current savings progress and target goals.

- **Savings Feature:**
    - **Detail Savings Page (`DetailSavingsPage`):**
        - Creates a new page to display detailed progress for a specific savings goal.
        - Passes the selected `CategoryEntity` to the detail page via `go_router`.
        - Displays the goal amount, current saved amount, and progress percentage.
        - Implements a `GoalProgressBar` to visually track progress.
        - Fetches and lists all transactions associated with the savings category.
        - Adds an "Add Saving" button to log new contributions.
    - **State Management (`DetailSavingsCubit`):**
        - Manages state for the detail page, including loading status, transaction list, current amount, and target amount.
        - Fetches all necessary data from the `TransactionRepository` and `SummaryRepository`.

- **UI & Widget Enhancements:**
    - **`GoalProgressBar`:**
        - Refactors the progress bar to handle various percentage levels more gracefully.
        - The percentage text is now positioned dynamically to remain visible even at low progress levels.
        - The target amount's text color changes for better contrast when the progress bar is nearly full.
    - **`TypeTransaction` Enum:**
        - Extends the enum with helper methods (`formatAmount`, `amountTextStyle`) for consistent formatting of transaction amounts and styles across the UI.
    - **`SimpleTransactionItem`:**
        - Uses the new `TypeTransaction` extension methods to simplify amount display and styling.
    - **`SummaryCard`:**
        - Formats the displayed amount to two decimal places.
    - **Layout Adjustments:**
        - Adjusts header heights and spacing on `HomePage`, `AnalysisPage`, `TransactionPage`, and `CategoriesPage` for a more consistent layout.
        - Disables horizontal swiping on the main page's `PageView`.

+++ lib/ui/widgets/button/target_button.dart
@@ -3,7 +3,7 @@ import 'package:finance_management/common/app_text_style.dart';
 import 'package:flutter/material.dart';
 
 class TargetButton extends StatelessWidget {
-  final int percent;
+  final double percent;
   final String targetName;
   final double minHeigh;
   final double minWidth;
